Add the option for the rotten apples
In this modification, the player can choose if he/she wants to add an
extra difficulty to the game. A part from the walls option, here the
appearance of rotten apples can also be chosen. If the snake eats one of
these rotten apples, its length will be reduced by one.
reduce the length of the snake by one.